[CHORE] UserSkillService 추상화
